Flytt aksjonsknapper for os og sp til egen komponent
